test(deposit): add theme test
