+++ app/components/UI/Ramp/Deposit/utils/utils.test.ts
@@ -6,6 +6,7 @@ import {
   getNotificationDetails,
   formatCurrency,
   hasDepositOrderField,
+  generateThemeParameters,
 } from '.';
 import { FiatOrder } from '../../../../../reducers/fiatOrders';
 import {
@@ -16,6 +17,8 @@ import { DepositOrder, DepositOrderType } from '@consensys/native-ramps-sdk';
 import { strings } from '../../../../../../locales/i18n';
 import { DepositPaymentMethod } from '../constants';
 import { IconName } from '../../../../../component-library/components/Icons/Icon';
+import { darkTheme, lightTheme } from '@metamask/design-tokens';
+import { AppThemeKey } from '../../../../../util/theme/models';
 
 jest.mock('../../../../../../locales/i18n', () => ({
   strings: jest.fn(),
@@ -345,5 +348,61 @@ describe('hasDepositOrderField', () => {
     expect(hasDepositOrderField(validDepositOrder, 'fiatAmount')).toBe(true);
     expect(hasDepositOrderField(validDepositOrder, 'network')).toBe(true);
     expect(hasDepositOrderField(validDepositOrder, 'status')).toBe(true);
+  });
+});
+
+describe.only('generateThemeParameters', () => {
+  it('should generate correct theme parameters for light mode', () => {
+    const themeAppearance = AppThemeKey.light;
+    const colors = lightTheme.colors;
+    const result = generateThemeParameters(themeAppearance, colors);
+    expect(result).toEqual({
+      themeColor: encodeURIComponent(colors.icon.default),
+      colorMode: 'LIGHT',
+      backgroundColors: [
+        colors.background.default,
+        colors.background.alternative,
+        colors.background.muted,
+      ]
+        .map(encodeURIComponent)
+        .join(','),
+      textColors: [colors.text.default, colors.text.default, colors.text.muted]
+        .map(encodeURIComponent)
+        .join(','),
+      borderColors: [
+        colors.border.default,
+        colors.border.muted,
+        colors.border.muted,
+      ]
+        .map(encodeURIComponent)
+        .join(','),
+    });
+  });
+
+  it('should generate correct theme parameters for dark mode', () => {
+    const themeAppearance = AppThemeKey.dark;
+    const colors = darkTheme.colors;
+    const result = generateThemeParameters(themeAppearance, colors);
+    expect(result).toEqual({
+      themeColor: encodeURIComponent(colors.icon.default),
+      colorMode: 'DARK',
+      backgroundColors: [
+        colors.background.default,
+        colors.background.alternative,
+        colors.background.muted,
+      ]
+        .map(encodeURIComponent)
+        .join(','),
+      textColors: [colors.text.default, colors.text.default, colors.text.muted]
+        .map(encodeURIComponent)
+        .join(','),
+      borderColors: [
+        colors.border.default,
+        colors.border.muted,
+        colors.border.muted,
+      ]
+        .map(encodeURIComponent)
+        .join(','),
+    });
   });
 });
